MH-4658, MH-4692, MH-2534, MH-4507: Scheduler submission bugs fixed. Delete, View, Edit links added to recording pages.
git-svn-id: https://opencast.jira.com/svn/MH/trunk@8048 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ modules/matterhorn-admin-ui/src/main/resources/ui/js/recordings.js
@@ -367,15 +367,17 @@ Recordings.removeRecording = function(workflowId) {
   });
 }
 
-Recordings.removeSchedulerEvent = function(eventId) {
-$.ajax({
-    url        : '../scheduler/rest/event/'+eventId,
-    type       : 'DELETE',
-    error      : function(XHR,status,e){
-      alert('Could not remove Scheduler Event ' + workflowId);
-    },
-    success    : function(data) {
-      location.reload();
-    }
-  });
+Recordings.removeScheduledRecording = function(eventId, title) {
+  if(confirm('Are you sure you wish to delete ' + title + '?')){
+    $.ajax({
+        url        : '/scheduler/rest/event/'+eventId,
+        type       : 'DELETE',
+        error      : function(XHR,status,e){
+          alert('Could not remove Scheduler Event ' + workflowId);
+        },
+        success    : function(data) {
+          location.reload();
+        }
+      });
+  }
 }
